refactor(server): move GoogleAuthStatus enum to LensmaniaLibrary.Enums
